+++ constraints.go
@@ -24,7 +24,7 @@ type Ordered interface {
 
 // Numeric is the set of integer and floating-point types.
 type Numeric interface {
-	~int | ~int8 | ~int16 | ~int32 | ~int64 | ~uint | ~uint8 | ~uint16 | ~uint32 | ~uint64 | ~float32 | ~float64
+	~int | ~int8 | ~int16 | ~int32 | ~int64 | ~uint | ~uint8 | ~uint16 | ~uint32 | ~uint64 | ~uintptr | ~float32 | ~float64
 }
 
 // Iterable documents that a value is expected to be a string, array, slice, map or channel.